Get rid of most pytest warnings.
Register all our marks.

Move a few test utilities that had been placed in the wrong utils.py file (which is another reason I moved the test one to test_utils.py).
