+++ pytest.ini
@@ -1,2 +1,13 @@
 [pytest]
-addopts = -rs --cov flask_security --cov-report term-missing --cache-clear
+addopts = -rs --cov flask_security --cov-report term-missing --cache-clear --strict-markers
+markers =
+    settings
+    babel
+    changeable
+    confirmable
+    registerable
+    two_factor
+    recoverable
+    passwordless
+    trackable
+    unified_signin
