+++ tests/test_confirmable.py
@@ -14,9 +14,13 @@ from flask import Flask
 from flask_security.core import UserMixin
 from flask_security.confirmable import generate_confirmation_token
 from flask_security.signals import confirm_instructions_sent, user_confirmed
-from flask_security.utils import capture_flashes, capture_registrations
 
-from tests.test_utils import authenticate, logout
+from tests.test_utils import (
+    authenticate,
+    capture_flashes,
+    capture_registrations,
+    logout,
+)
 
 pytestmark = pytest.mark.confirmable()
