+++ tests/test_utils.py
@@ -7,13 +7,20 @@
     :copyright: (c) 2019-2020 by J. Christopher Wagner (jwag).
     :license: MIT, see LICENSE for more details.
 """
+from contextlib import contextmanager
 
 from flask.json.tag import TaggedJSONSerializer
+from flask.signals import message_flashed
 
 from flask_security import Security, SmsSenderBaseClass
 from flask_security.datastore import (
     SQLAlchemyUserDatastore,
     SQLAlchemySessionUserDatastore,
+)
+from flask_security.signals import (
+    login_instructions_sent,
+    reset_password_instructions_sent,
+    user_registered,
 )
 from flask_security.utils import hash_password
 
@@ -199,3 +206,71 @@ class SmsTestSender(SmsSenderBaseClass):
 class SmsBadSender(SmsSenderBaseClass):
     def send_sms(self, from_number, to_number, msg):
         raise ValueError(f"Unknown number: {to_number}")
+
+
+@contextmanager
+def capture_passwordless_login_requests():
+    login_requests = []
+
+    def _on(app, **data):
+        login_requests.append(data)
+
+    login_instructions_sent.connect(_on)
+
+    try:
+        yield login_requests
+    finally:
+        login_instructions_sent.disconnect(_on)
+
+
+@contextmanager
+def capture_registrations():
+    """Testing utility for capturing registrations.
+    """
+    registrations = []
+
+    def _on(app, **data):
+        registrations.append(data)
+
+    user_registered.connect(_on)
+
+    try:
+        yield registrations
+    finally:
+        user_registered.disconnect(_on)
+
+
+@contextmanager
+def capture_reset_password_requests(reset_password_sent_at=None):
+    """Testing utility for capturing password reset requests.
+
+    :param reset_password_sent_at: An optional datetime object to set the
+                                   user's `reset_password_sent_at` to
+    """
+    reset_requests = []
+
+    def _on(app, **data):
+        reset_requests.append(data)
+
+    reset_password_instructions_sent.connect(_on)
+
+    try:
+        yield reset_requests
+    finally:
+        reset_password_instructions_sent.disconnect(_on)
+
+
+@contextmanager
+def capture_flashes():
+    """Testing utility for capturing flashes."""
+    flashes = []
+
+    def _on(app, **data):
+        flashes.append(data)
+
+    message_flashed.connect(_on)
+
+    try:
+        yield flashes
+    finally:
+        message_flashed.disconnect(_on)
