+++ tests/view_scaffold.py
@@ -183,9 +183,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
-engine = create_engine(
-    "sqlite:////tmp/view_scaffold.db?check_same_thread=False", convert_unicode=True
-)
+engine = create_engine("sqlite:////tmp/view_scaffold.db?check_same_thread=False")
 db_session = scoped_session(
     sessionmaker(autocommit=False, autoflush=False, bind=engine)
 )
